bot/requirements.txt
aiohttp~=3.9.1
uvicorn~=0.24.0
fastapi~=0.104.1
pydantic~=2.5.0
python-telegram-bot~=20.7